matern kernel is possible to use in MMD

+++ tests/test_backend_torch.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from model_criticism_mmd.backends.backend_torch import TypeInputData, ModelTrainerTorchBackend
 from model_criticism_mmd.logger_unit import logger
 from model_criticism_mmd.supports import distribution_generator
-from sklearn.utils import check_random_state
+from model_criticism_mmd.backends.kernels_torch import RBFKernelFunction, MaternKernelFunction
 
 import torch
 
@@ -24,24 +24,24 @@ def test_devel(resource_path_root: Path):
     x_test = array_obj['x_test']
     y_test = array_obj['y_test']
     init_scale = torch.tensor(np.array([0.05, 0.55]))
+    for kernel_function in [MaternKernelFunction(nu=0.5), RBFKernelFunction()]:
+        trainer = ModelTrainerTorchBackend(kernel_function_obj=kernel_function)
 
-    trainer = ModelTrainerTorchBackend()
-
-    trained_obj = trainer.train(x_train,
-                                y_train,
-                                num_epochs=num_epochs,
-                                x_val=x_test,
-                                y_val=y_test,
-                                initial_scale=init_scale,
-                                opt_sigma=True, opt_log=True, init_sigma_median=False)
-    trained_obj.to_pickle(path_trained_model)
-    import math
-    logger.info(f'exp(sigma)={math.exp(trained_obj.sigma)} scales={trained_obj.scales}')
-
-    assert np.linalg.norm(trained_obj.scales[0] - trained_obj.scales[1]) < 5
-    assert 0.0 < math.exp(trained_obj.sigma) < 1.5
-    os.remove(path_trained_model)
-
+        trained_obj = trainer.train(x_train,
+                                    y_train,
+                                    num_epochs=num_epochs,
+                                    x_val=x_test,
+                                    y_val=y_test,
+                                    initial_scale=init_scale,
+                                    opt_sigma=True, opt_log=True, init_sigma_median=False)
+        trained_obj.to_pickle(path_trained_model)
+        import math
+        logger.info(f'exp(sigma)={math.exp(trained_obj.sigma)} scales={trained_obj.scales}')
+        if isinstance(kernel_function, RBFKernelFunction):
+            assert np.linalg.norm(trained_obj.scales[0] - trained_obj.scales[1]) < 5
+            assert 0.0 < math.exp(trained_obj.sigma) < 1.5
+            os.remove(path_trained_model)
+        # end if
 
 def test_example():
     n_train = 1500
